Move component, migrate to audio buffers
